Changed the type to 'size_t'.

+++ src/game_objects/tile_map/tile_map.cpp
@@ -2,7 +2,7 @@
 
 TileMap::TileMap(Vector2 position, unsigned int tileSize) : GeneralGameObject(position), tileSize(tileSize)
 {
-    for (unsigned int i = 0; i < NUMBER_OF_TILES; i++)
+    for (size_t i = 0; i < NUMBER_OF_TILES; i++)
     {
         TILE_MAP_RECTANGLES_POSITION
         this->rectangles[i].w = tileSize;
@@ -11,7 +11,7 @@ TileMap::TileMap(Vector2 position, unsigned int tileSize) : GeneralGameObject(po
 }
 void TileMap::update(double deltaTime)
 {
-    for (unsigned int i = 0; i < NUMBER_OF_TILES; i++)
+    for (size_t i = 0; i < NUMBER_OF_TILES; i++)
     {
         TILE_MAP_RECTANGLES_POSITION
     }
@@ -19,7 +19,7 @@ void TileMap::update(double deltaTime)
 void TileMap::draw(void)
 {
     // SDL_SetRenderDrawColour(renderer, this->colour.r, this->colour.g, this->colour.b, this->colour.a); // Sets draw colour.
-    for (unsigned int i = 0; i < NUMBER_OF_TILES; i++)
+    for (size_t i = 0; i < NUMBER_OF_TILES; i++)
     {
         // if (i % 3 == 0)
         //     SDL_SetRenderDrawColour(renderer, 0XFF, 0X33, 0X33, 0X00);
@@ -38,7 +38,7 @@ void TileMap::draw(void)
     GeneralGameObject::draw();
 }
 
-void TileMap::setTile(unsigned int x, unsigned int y, TILE_TYPE tileType)
+void TileMap::setTile(size_t x, size_t y, TILE_TYPE tileType)
 {
     this->tileMap[y * WIDTH_OF_TILE_MAP + x % WIDTH_OF_TILE_MAP] = tileType;
 }
